Move agentd's middlewares to methods of agentd
